refactored names and packages to follow rules.
