+++ .idea/workspace.xml
@@ -1,21 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="ChangeListManager">
-    <list default="true" id="0f1f7fa8-e9e9-4070-b28f-5cbc42e87279" name="Changes" comment="new howard branch used to finish the todos ansh made">
+    <list default="true" id="0f1f7fa8-e9e9-4070-b28f-5cbc42e87279" name="Changes" comment="refactored names and packages to follow rules.">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/out/production/1159 pm/save.ser" beforeDir="false" afterPath="$PROJECT_DIR$/out/production/1159 pm/save.ser" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/out/test/1159 pm/EventTest.class" beforeDir="false" afterPath="$PROJECT_DIR$/out/test/1159 pm/EventTest.class" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/out/test/1159 pm/MementoTest.class" beforeDir="false" afterPath="$PROJECT_DIR$/out/test/1159 pm/MementoTest.class" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/out/test/1159 pm/ProvinceTest.class" beforeDir="false" afterPath="$PROJECT_DIR$/out/test/1159 pm/ProvinceTest.class" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/gamemanager/Decisions.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/gamemanager/Decisions.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/gamemanager/GameEngine.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/gamemanager/GameEngine.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/gamesave/GameStateElement.java" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/ui/UserInterface.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/ui/UserInterface.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/save.ser" beforeDir="false" afterPath="$PROJECT_DIR$/src/save.ser" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/test/DecisionTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/test/DecisionTest.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/test/EventTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/test/EventTest.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/test/MementoTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/test/MementoTest.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/test/ProvinceTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/test/ProvinceTest.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -121,20 +108,24 @@
     <property name="project.structure.side.proportion" value="0.2" />
     <property name="run.code.analysis.last.selected.profile" value="pProject Default" />
   </component>
-  <component name="RunManager" selected="JUnit.ProvinceTest">
-    <configuration name="All in 1159 pm" type="JUnit" factoryName="JUnit" temporary="true" nameIsGenerated="true">
+  <component name="RunManager" selected="Application.Main">
+    <configuration name="Main" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
+      <option name="MAIN_CLASS_NAME" value="main.java.Main" />
       <module name="1159 pm" />
-      <option name="PACKAGE_NAME" value="" />
-      <option name="TEST_OBJECT" value="package" />
+      <extension name="coverage">
+        <pattern>
+          <option name="PATTERN" value="main.java.*" />
+          <option name="ENABLED" value="true" />
+        </pattern>
+      </extension>
       <method v="2">
         <option name="Make" enabled="true" />
       </method>
     </configuration>
-    <configuration name="GameEngineTest" type="JUnit" factoryName="JUnit" temporary="true" nameIsGenerated="true">
+    <configuration name="All in 1159 pm" type="JUnit" factoryName="JUnit" temporary="true" nameIsGenerated="true">
       <module name="1159 pm" />
       <option name="PACKAGE_NAME" value="" />
-      <option name="MAIN_CLASS_NAME" value="GameEngineTest" />
-      <option name="TEST_OBJECT" value="class" />
+      <option name="TEST_OBJECT" value="package" />
       <method v="2">
         <option name="Make" enabled="true" />
       </method>
@@ -171,10 +162,10 @@
     <recent_temporary>
       <list>
         <item itemvalue="JUnit.ProvinceTest" />
-        <item itemvalue="JUnit.All in 1159 pm" />
-        <item itemvalue="JUnit.GameEngineTest.testDeath" />
         <item itemvalue="JUnit.GameEngineTest.testProvinceDeath" />
-        <item itemvalue="JUnit.GameEngineTest" />
+        <item itemvalue="JUnit.GameEngineTest.testDeath" />
+        <item itemvalue="JUnit.All in 1159 pm" />
+        <item itemvalue="Application.Main" />
       </list>
     </recent_temporary>
   </component>
@@ -283,7 +274,14 @@
       <option name="project" value="LOCAL" />
       <updated>1638758462293</updated>
     </task>
-    <option name="localTasksCounter" value="12" />
+    <task id="LOCAL-00012" summary="refactored names and packages to follow rules.">
+      <created>1638929103981</created>
+      <option name="number" value="00012" />
+      <option name="presentableId" value="LOCAL-00012" />
+      <option name="project" value="LOCAL" />
+      <updated>1638929103981</updated>
+    </task>
+    <option name="localTasksCounter" value="13" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -308,7 +306,8 @@
     <MESSAGE value="i literally did everything" />
     <MESSAGE value="new howard branch used to finish the todos ansh made" />
     <MESSAGE value="pls work" />
-    <option name="LAST_COMMIT_MESSAGE" value="pls work" />
+    <MESSAGE value="refactored names and packages to follow rules." />
+    <option name="LAST_COMMIT_MESSAGE" value="refactored names and packages to follow rules." />
   </component>
   <component name="XDebuggerManager">
     <breakpoint-manager>
@@ -322,8 +321,8 @@
     </breakpoint-manager>
   </component>
   <component name="com.intellij.coverage.CoverageDataManagerImpl">
-    <SUITE FILE_PATH="coverage/course_project$MementoTest.ic" NAME="MementoTest Coverage Results" MODIFIED="1636989429253" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="idea" COVERAGE_BY_TEST_ENABLED="false" COVERAGE_TRACING_ENABLED="false" />
     <SUITE FILE_PATH="coverage/course_project$GameManagerTest_testProcessEvent.ic" NAME="GameManagerTest.testProcessEvent Coverage Results" MODIFIED="1638413855525" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="idea" COVERAGE_BY_TEST_ENABLED="false" COVERAGE_TRACING_ENABLED="false" />
     <SUITE FILE_PATH="coverage/course_project_11_59pmver5$All_in_1159_pm.ic" NAME="All in 1159 pm Coverage Results" MODIFIED="1638782346221" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="idea" COVERAGE_BY_TEST_ENABLED="false" COVERAGE_TRACING_ENABLED="false" />
+    <SUITE FILE_PATH="coverage/course_project$MementoTest.ic" NAME="MementoTest Coverage Results" MODIFIED="1636989429253" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="idea" COVERAGE_BY_TEST_ENABLED="false" COVERAGE_TRACING_ENABLED="false" />
   </component>
 </project>
