+++ src/main/java/usecases/ProcessValues.java
@@ -1,6 +1,6 @@
 package main.java.usecases;
 
-import main.java.provinceconstruction.Province;
+import main.java.province_construction.Province;
 
 import java.util.List;
