+++ test/BattleTest.java
@@ -5,9 +5,9 @@
   */
 import main.java.usecases.Battle;
 
-import main.java.provinceconstruction.ProvinceAssembler;
-import main.java.provinceconstruction.Province;
-import main.java.provinceconstruction.ProvinceBuilder;
+import main.java.province_construction.ProvinceAssembler;
+import main.java.province_construction.Province;
+import main.java.province_construction.ProvinceBuilder;
 import org.junit.*;
 
 import java.util.*;
